pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalrun"
version = "0.1.0"
description = "Objective LLM Evaluation Suite"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "anthropic>=0.39",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "tabulate>=0.9",
]

[project.scripts]
evalrun = "evalrun.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
where = ["src"]
